Mostly works but needs some tuning with turning especially.

+++ starstruck_dumper/include/Utilities.h
@@ -14,7 +14,7 @@
 int limit(int, int, int);
 int abs(int);
 int inDeadBand(int realNum, int desiredNum, int deadBand);
-double encoderToInches(int encoderLines, double wheelDiameter);
+double encoderToInches(double encoderLines, double wheelDiameter);
 double absDouble(double num);
 int inDeadBand(int realNum, int desiredNum, int deadBand);
 int inDeadBandDouble(double realNum, double desiredNum, double deadBand);
