+++ starstruck_dumper/src/Utilities.c
@@ -34,7 +34,7 @@ int inDeadBand(int realNum, int desiredNum, int deadBand)
 	return (abs(realNum - desiredNum) < deadBand);
 }
 
-double encoderToInches(int encoderLines, double wheelDiameter)
+double encoderToInches(double encoderLines, double wheelDiameter)
 {
 	return ((encoderLines / 360.0) * PI * wheelDiameter);
 }
